Refactor InputLoopService into Program.cs for CosmosDB transactions sample
- Removed InputLoopService and moved logic back to Program.cs Main method
- Updated to use StartAsync/StopAsync pattern with IMessageSession extraction
- Maintains compatibility with hosting model while enabling proper console interaction
- Removed Console.ReadKey() from Server for containerization compatibility
