+++ samples/cosmosdb/transactions/CosmosDB_3/Client/Program.cs
@@ -1,36 +1,42 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Client;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
 
 class Program
 {
-    public static async Task Main(string[] args)
+    static async Task Main()
     {
-        await CreateHostBuilder(args).Build().RunAsync();
+        Console.Title = "Client";
+        var endpointConfiguration = new EndpointConfiguration("Samples.CosmosDB.Transactions.Client");
+        endpointConfiguration.UsePersistence<LearningPersistence>();
+        endpointConfiguration.UseTransport(new LearningTransport());
+        endpointConfiguration.UseSerialization<SystemJsonSerializer>();
+
+        var endpointInstance = await Endpoint.Start(endpointConfiguration);
+
+        Console.WriteLine("Press 'S' to send a StartOrder message to the server endpoint");
+        Console.WriteLine("Press any other key to exit");
+
+        while (true)
+        {
+            var key = Console.ReadKey();
+            Console.WriteLine();
+
+            var orderId = Guid.NewGuid();
+            var startOrder = new StartOrder
+            {
+                OrderId = orderId
+            };
+            if (key.Key == ConsoleKey.S)
+            {
+                await endpointInstance.Send("Samples.CosmosDB.Transactions.Server", startOrder);
+                Console.WriteLine($"StartOrder Message sent to Server with OrderId {orderId}");
+                continue;
+            }
+            break;
+        }
+
+        await endpointInstance.Stop();
     }
-
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-     Host.CreateDefaultBuilder(args)
-         .ConfigureServices((hostContext, services) =>
-         {
-             Console.Title = "Server";
-             services.AddHostedService<InputLoopService>();
-
-         }).UseNServiceBus(x =>
-         {
-             Console.Title = "Client";
-             var endpointConfiguration = new EndpointConfiguration("Samples.CosmosDB.Transactions.Client");
-             endpointConfiguration.UsePersistence<LearningPersistence>();
-             endpointConfiguration.UseTransport(new LearningTransport());
-             endpointConfiguration.UseSerialization<SystemJsonSerializer>();
-
-             Console.WriteLine("Press 'S' to send a StartOrder message to the server endpoint");
-
-             Console.WriteLine("Press any other key to exit");
-             return endpointConfiguration;
-         });
-
 }
